fix: more accurate parsing of sentences
We no longer split sentences based on `.`, but on `. `. The extra bit of
accuracy prevents us from getting definitions that break in the middle
of a sentence like `The year was 11.33.55 when the martians invaded.`

However, there are still limitations around acronyms since many of them
end with `. ` (eg., a.k.a., etc.).