feat: handle wikitext indents
Indent was a wikitext rule that was being missed. This was caught because the dictfile parser doesn't like the pattern `\n:` as it denotes "word info". This was causing issues when trying to convert kobo dictionaries. We now handle indents appropriately, cleaning the colon and instead adding two spaces per colon. This also lets us handle nesting.